+++ src/main/java/net/frozenblock/freezeframe/mixin/client/camera/LevelRendererMixin.java
@@ -18,12 +18,15 @@
 package net.frozenblock.freezeframe.mixin.client.camera;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
 import com.llamalad7.mixinextras.sugar.Local;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.frozenblock.freezeframe.client.screenshot.FFScreenshotUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.LevelRenderer;
+import net.minecraft.client.renderer.culling.Frustum;
+import net.minecraft.client.renderer.debug.DebugRenderer;
 import net.minecraft.world.entity.Entity;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -54,5 +57,16 @@ public class LevelRendererMixin {
 		if (FFScreenshotUtil.notScreenshottingOrIsTripod()) return original;
 
 		return entity == this.minecraft.player ? entity : original;
+	}
+
+	@WrapWithCondition(
+		method = "extractLevel",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/client/renderer/debug/DebugRenderer;emitGizmos(Lnet/minecraft/client/renderer/culling/Frustum;DDDF)V"
+		)
+	)
+	public boolean freezeFrame$ignoreGizmosWhileScreenshotting(DebugRenderer instance, Frustum frustum, double camX, double camY, double camZ, float partialTicks) {
+		return !FFScreenshotUtil.screenshotting();
 	}
 }